feat(lotto): 수익률 계산 기능 구현
- 수익률(단위: %) = (총 상금 / 구입 금액) * 100
- 총 상금 = 각 등수의 (상금 * 당첨 개수)를 모두 합한 값

+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class LottoGenerator {
-    public List<Lotto> generateLotto(int howManyLotto) {
+    public List<Lotto> generateLotto(long howManyLotto) {
         List<Lotto> lottos = new ArrayList<>();
         for (int i = 0; i < howManyLotto; i++) {
             Lotto lotto = new Lotto(generateNumbers());
